+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Changelog for NeoFS Node
 - `neofs-cli` does not allow searches based on homomorphic hash values (#3847)
 - `neofs-cli`, `neofs-node` do not allow searches based on homomorphic hash values (#3847)
 - Storage nodes do not calculate homomorphic hashes for objects (#3847)
+- Optimized GET/HEAD request forwarding (#3877)
 
 ### Removed
 - `policer.max_workers` configuration (#3920)